// backend-cm/prisma/investigation/generated/sanction_status_code.ts
import { administrative_sanction } from "./administrative_sanction";
import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";

export class sanction_status_code {
  @ApiProperty({ type: String })
  sanction_status_code: string;

  @ApiProperty({ type: String })
  short_description: string;

  @ApiPropertyOptional({ type: String })
  long_description?: string;

  @ApiPropertyOptional({ type: Number })
  display_order?: number;

  @ApiProperty({ type: Boolean })
  active_ind: boolean = true;

  @ApiProperty({ type: String })
  create_user_id: string;

  @ApiProperty({ type: Date })
  create_utc_timestamp: Date;

  @ApiPropertyOptional({ type: String })
  update_user_id?: string;

  @ApiPropertyOptional({ type: Date })
  update_utc_timestamp?: Date;

  @ApiProperty({ isArray: true, type: () => administrative_sanction })
  administrative_sanction_administrative_sanction_sanction_status_codeTosanction_status_code: administrative_sanction[];
}